delete

+++ delete.php
@@ -1,7 +1,10 @@
 <?php
-    $sabor=$_GET['codigo_s'];
-    $db=new SQLite3('pizza.db');
-    $db->query("delete from sabor where sabor.codigo=$sabor");
+    if(isset($_GET['codigo_s'])){
+        $sabor=$_GET['codigo_s'];
+        $db=new SQLite3('pizza.db');
+        $db->query("delete from sabor where sabor.codigo=$sabor");
+    }
     header( "refresh:1;url=index.php" );
-    die();   
+    die();
+       
 ?>
